move bundle to spiffe
